[Fix#49] 자산 설정 타입 버그 수정
int -> long 으로 타입 변경

+++ data/src/main/java/com/aos/data/repository/remote/book/BookRepositoryImpl.kt
@@ -484,7 +484,7 @@ class BookRepositoryImpl @Inject constructor(private val bookDataSource: BookRem
             }
         }
     }
-    override suspend fun postBooksInfoAsset(bookKey: String, asset: Int): Result<Void?> {
+    override suspend fun postBooksInfoAsset(bookKey: String, asset: Long): Result<Void?> {
         when (val data =
             bookDataSource.postBooksInfoAsset(PostBooksInfoAssetBody(bookKey, asset))) {
             is NetworkState.Success -> {
