+++ domain/src/main/java/com/aos/repository/BookRepository.kt
@@ -122,7 +122,7 @@ interface BookRepository {
     suspend fun getBooksInfoCurrency(bookKey: String): Result<GetBooksInfoCurrencyModel>
 
     // 가계부 자산 설정하기
-    suspend fun postBooksInfoAsset(bookKey: String, asset: Int): Result<Void?>
+    suspend fun postBooksInfoAsset(bookKey: String, asset: Long): Result<Void?>
 
     // 가계부 이월설정 On/Off
     suspend fun postBooksInfoCarryOver(status: Boolean, bookKey: String): Result<Void?>
